genric report excel report sum and pdf report sum not matching fix changed the computed value to decimal

+++ ENTITY/Reports/ATTShareHolderReportTotalBasedOn.cs
@@ -4,7 +4,7 @@
     {
         public string TotalBasedOn { get; set; }
         public string Location { get; set; }
-        public float CalculateTotal { get; set; }
+        public decimal CalculateTotal { get; set; }
         public int TotalIndex { get; set; }
 
     }
